docker/services/execution/requirements.txt
cryptography==35.0.0
eventlet==0.30.2
flask==2.0.1
flask-jwt-extended==4.3.0
flask-migrate==2.4.0
flask-script==2.0.6
Flask-SQLAlchemy==2.3.2
flask-talisman==1.0.0
flask_cors==3.0.7
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.18.2
pandas==1.0.4
plotly==5.8.0
psutil==5.8.0
psycopg2-binary==2.8.5
pyjwt==2.4.0
sentry-sdk[flask]==1.24.0
Werkzeug==2.0.3
